requirements.txt
click==8.1.3
Flask==3.0
Flask-SQLAlchemy==3.1
gunicorn==21.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.1
Werkzeug==3.0
